Get database more or less working well enough to be considered a database

+++ music-theori/Database/IHasPrimaryKey.cs
@@ -2,6 +2,6 @@
 {
     public interface IHasPrimaryKey
     {
-        int ID { get; set; }
+        long ID { get; set; }
     }
 }
